Implemented the reformating itself; processing works with f64 now

+++ reformat/src/core/alignment_loader.rs
@@ -51,7 +51,7 @@ pub(crate) struct Row {
     /// Reference sequence name this read aligns to (if applicable)
     ref_region: Option<ReferenceRegion>,
     /// Z-standardized raw current measurements
-    signal: Vec<f32>
+    signal: Vec<f64>
 }
 
 
@@ -69,7 +69,7 @@ impl Row {
         read_id: Uuid,
         alignment: Vec<usize>,
         sequence: Vec<u8>,
-        signal: Vec<f32>,
+        signal: Vec<f64>,
         ref_name: Option<String>,
         ref_start: Option<usize> 
     ) -> Result<Self, RowError> {
@@ -112,7 +112,7 @@ impl Row {
     }
 
     /// Returns the raw signal data.
-    pub(crate) fn signal(&self) -> &[f32] {
+    pub(crate) fn signal(&self) -> &[f64] {
         &self.signal
     }
 }
